Update #38 Progress being made, but not quite yet!
So we now finally have a renderable "effect", it appears to be the Mesh Oriented at the default angle, with no rotation ( probably drawn 6 times, indicating no face rotation is working ). Texture is wrong, but its likely easy to fix

It also appears the block is not properly being saved/reloaded from file, and work needs to be made towards DebugDesignerBlock before its "Working" "sucessfully"

+++ src/com/GlitchyDev/Utility/FrustumCullingFilter.java
@@ -6,7 +6,6 @@ import com.GlitchyDev.World.Region.Region;
 import org.joml.FrustumIntersection;
 import org.joml.Matrix4f;
 import org.joml.Vector3f;
-import org.joml.Vector3i;
 
 import java.util.Collection;
 import java.util.List;
@@ -40,7 +39,7 @@ public class FrustumCullingFilter {
 
     public static void filter(Collection<Region> regions) {
         float boundingRadius;
-        Vector3i pos;
+        Vector3f pos;
         for(Region region: regions) {
             for(Block block: region.getBlocksArray()) {
                 boundingRadius = 1.80f;
